Change a whole host of framework configurations
There are a lot of changes here, from volte support, which probably
does nothing at all, to doze support (which will require gapps or other
cloud support) to long-hold and double-tap of the home button to activate
search and menu to auto-on buttons etc.

This is meant to be tweaked further, obviously, but didn't seem to
result in anything "bad".  Let's get some response and we can start
adjusting further.

The stylus gestures menu is now available, however for some reason to
be investigated, the gestures are not recognized.  This may be because
the stylus button press (BTN_STYLUS) is not properly becoming a framework
event.  The wacom driver is definitely detecting the press-- it remains
to be seen whether this is propogating to Android or not..  You can see
the detection of the button press by hitting the stylus button and checking
dmesg.

Change-Id: I8c8d2c6707dbf1008dda80b7ed7248728170755c